+++ packages/cookie-parser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinyhttp/cookie-parser
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/cookie@0.2.1
+  - @tinyhttp/app@0.3.1
+
 ## 0.2.0
 
 ### Minor Changes