+++ packages/ip-filter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/ip-filter
 
+## 0.1.1
+
+### Patch Changes
+
+- @tinyhttp/app@0.3.1
+
 ## 0.1.0
 
 ### Minor Changes