+++ packages/res/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/res
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/cookie@0.2.1
+
 ## 0.1.0
 
 ### Minor Changes